Update Cloud API interfaces to use new Region API endpoints
The new Region API runs in each cloud region and replaces
the APIs currently embedded in the Region Controller and
Environment Controllers in each cloud region.

This updates the src/cloud-api crate to point to these new
Region API endpoints.

NOTE that this does change the naming of many of the Cloud API
client's public methods by re-aligning the naming to use the
customer-facing conventions (replace 'environment' with 'region')
