+++ src/cloud-api/Cargo.toml
@@ -9,6 +9,7 @@ publish = false
 
 [dependencies]
 anyhow = "1.0.44"
+chrono = { version = "0.4.23", default-features = false, features = ["std"] }
 reqwest = { version = "0.11.4", features = ["json"] }
 once_cell = "1.16.0"
 serde = { version = "1.0.130", features = ["derive"] }
