+++ src/cloud-api/src/client.rs
@@ -10,7 +10,7 @@
 //! # Materialize cloud API client
 //!
 //! This module provides an API client with typed methods for
-//! interacting with the Materialize cloud API. This client i,
+//! interacting with the Materialize cloud API. This client includes,
 //! token management, and basic requests against the API.
 //!
 //! The [`Client`] requires an [`mz_frontegg_client::client::Client`] as a parameter. The
@@ -21,7 +21,6 @@ use reqwest::{Method, RequestBuilder, Url};
 use serde::de::DeserializeOwned;
 use serde::Deserialize;
 
-use crate::client::region::Region;
 use crate::error::{ApiError, Error};
 
 use self::cloud_provider::CloudProvider;
@@ -34,19 +33,18 @@ pub struct Client {
 }
 
 pub mod cloud_provider;
-pub mod environment;
 pub mod region;
 
 /// Cloud endpoints architecture:
 ///
-/// (CloudProvider)                         (Region)                                 (Environment)
-///   ---------              --------------------------------------            ------------------------
-///  |          |           |          Region Controller           |          | Environment Controller |
-///  |  Cloud   |  api_url  |    ----------        -------------   |  ec_url  |                        |
-///  |  Sync    | --------> |   | Provider | ---- |    Region   |  | -------> |       Environment      |
-///  |          |           |   | (aws..)  |      |  (east-1..) |  |          |  (pgwire_address...)   |
-///  |          |           |    ----------        -------------   |          |                        |
-///   ----------             --------------------------------------            -----------------------
+/// (CloudProvider)                                (Region)
+///   ---------                     --------------------------------------
+///  |          |                  |              Region API              |
+///  |  Cloud   |        url       |    ----------        -------------   |
+///  |  Sync    | ---------------> |   | Provider | ---- |    Region   |  |
+///  |          |                  |   | (aws..)  |      |  (east-1..) |  |
+///  |          |                  |    ----------        -------------   |
+///   ----------                    --------------------------------------
 ///
 impl Client {
     /// Builds a request towards the `Client`'s endpoint
@@ -61,7 +59,7 @@ impl Client {
     {
         let mut endpoint = self.endpoint.clone();
         endpoint.set_host(Some(&format!(
-            "sync.{}",
+            "api.{}",
             self.endpoint
                 .domain()
                 .ok_or_else(|| Error::InvalidEndpointDomain)?
@@ -72,37 +70,19 @@ impl Client {
 
     /// Builds a request towards the `Client`'s endpoint
     /// The function requires a [CloudProvider] as parameter
-    /// since it contains the api url (region controller url)
+    /// since it contains the api url (Region API url)
     /// to interact with the region.
     async fn build_region_request<P>(
         &self,
         method: Method,
         path: P,
-        cloud_provider: CloudProvider,
+        cloud_provider: &CloudProvider,
     ) -> Result<RequestBuilder, Error>
     where
         P: IntoIterator,
         P::Item: AsRef<str>,
     {
-        self.build_request(method, path, cloud_provider.api_url)
-            .await
-    }
-
-    /// Builds a request towards the `Client`'s endpoint.
-    /// The function requires a [Region] as parameter
-    /// since it contains the environment controller url
-    /// to interact with the environment.
-    async fn build_environment_request<P>(
-        &self,
-        method: Method,
-        path: P,
-        region: Region,
-    ) -> Result<RequestBuilder, Error>
-    where
-        P: IntoIterator,
-        P::Item: AsRef<str>,
-    {
-        self.build_request(method, path, region.environment_controller_url)
+        self.build_request(method, path, cloud_provider.url.clone())
             .await
     }
